+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "crates/shielded-pool",
   "crates/component",
   "crates/ibc",
+  "crates/dao",
   # Legacy: move these directories into `crates/` heirarchy
   # as it becomes possible (no in flight changes)
   "decaf377-fmd",
