+++ app/src/action_handler/actions.rs
@@ -10,9 +10,6 @@ use penumbra_transaction::Action;
 use super::ActionHandler;
 use penumbra_component::ActionHandler as _;
 
-mod dao_deposit;
-mod dao_output;
-mod dao_spend;
 mod delegate;
 mod delegator_vote;
 mod position;
