--- app/src/dao/mod.rs
@@ -1,5 +1,0 @@
-// The DAO isn't really a "component" because it doesn't really execute anything by itself. It's
-// just a collection of state that is modified by DaoSpend and DaoDeposit actions.
-
-pub mod state_key;
-pub mod view;
