+++ app/src/lib.rs
@@ -10,7 +10,6 @@ pub use mock_client::MockClient;
 pub use temp_storage_ext::TempStorageExt;
 
 pub mod app;
-pub mod dao;
 pub mod dex;
 pub mod governance;
 pub mod stake;
