+++ crates/dao/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "penumbra-dao"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+component = ["penumbra-component", "penumbra-storage", "penumbra-proto/penumbra-storage", "penumbra-chain/component", "penumbra-shielded-pool/component"]
+default = ["component"]
+docsrs = []
+
+[dependencies]
+# Workspace dependencies
+penumbra-proto = { path = "../../proto", default-features = false }
+penumbra-storage = { path = "../../storage", optional = true }
+penumbra-component = { path = "../component", optional = true }
+penumbra-chain = { path = "../../chain", default-features = false }
+penumbra-crypto = { path = "../../crypto", default-features = false }
+penumbra-shielded-pool = { path = "../shielded-pool", default-features = false }
+
+# Crates.io deps
+ark-ff = { version = "0.3", default_features = false }
+async-trait = "0.1.52"
+hex = "0.4"
+anyhow = "1"
+tracing = "0.1"
+prost = "0.11"
+serde = { version = "1", features = ["derive"] }
+metrics = "0.19.0"
+pbjson-types = "0.5.1"
+tendermint = "0.29.0"
+tendermint-light-client-verifier = "0.29.0"
+sha2 = "0.10.6"
+once_cell = "1.17.1"
+base64 = "0.20"
+blake2b_simd = "0.5"
+futures = "0.3.28"
+
+[dev-dependencies]
+tokio = {version = "1.3", features = ["full"]}
