+++ crates/dao/src/action.rs
@@ -1,0 +1,7 @@
+mod dao_deposit;
+mod dao_output;
+mod dao_spend;
+
+pub use dao_deposit::DaoDeposit;
+pub use dao_output::DaoOutput;
+pub use dao_spend::DaoSpend;
