+++ crates/dao/src/action/dao_spend.rs
@@ -1,27 +1,24 @@
 use std::convert::{TryFrom, TryInto};
 
 use anyhow::{Context, Error};
-use ark_ff::Zero;
-use decaf377::Fr;
-use penumbra_crypto::{balance, Balance, Value};
+use penumbra_crypto::{Balance, EffectHash, EffectingData, Value};
 use penumbra_proto::{core::governance::v1alpha1 as pb, DomainType};
-
-use crate::{ActionView, TransactionPerspective};
-
-use super::IsAction;
 
 #[derive(Clone, Debug)]
 pub struct DaoSpend {
     pub value: Value,
 }
 
-impl IsAction for DaoSpend {
-    fn balance_commitment(&self) -> balance::Commitment {
-        self.balance().commit(Fr::zero())
-    }
+impl EffectingData for DaoSpend {
+    fn effect_hash(&self) -> EffectHash {
+        let mut state = blake2b_simd::Params::default()
+            .personal(b"PAH:daospend")
+            .to_state();
 
-    fn view_from_perspective(&self, _txp: &TransactionPerspective) -> ActionView {
-        ActionView::DaoSpend(self.clone())
+        state.update(&self.value.amount.to_le_bytes());
+        state.update(&self.value.asset_id.to_bytes());
+
+        EffectHash(state.finalize().as_array().clone())
     }
 }
