+++ crates/dao/src/component/action_handler.rs
@@ -1,0 +1,3 @@
+mod dao_deposit;
+mod dao_output;
+mod dao_spend;
