+++ crates/dao/src/event.rs
@@ -1,0 +1,5 @@
+// TODO: events for
+//
+// - dao spend
+// - dao deposit
+// ?
