+++ crates/dao/src/lib.rs
@@ -1,0 +1,10 @@
+#![cfg_attr(docsrs, feature(doc_cfg))]
+
+#[cfg_attr(docsrs, doc(cfg(feature = "component")))]
+#[cfg(feature = "component")]
+pub mod component;
+
+pub mod event;
+
+mod action;
+pub use action::{DaoDeposit, DaoOutput, DaoSpend};
