+++ deployments/scripts/rust-docs
@@ -40,6 +40,7 @@ cargo +nightly doc --no-deps \
   -p penumbra-chain \
   -p penumbra-component \
   -p penumbra-app \
+  -p penumbra-dao \
   -p penumbra-crypto \
   -p penumbra-custody \
   -p penumbra-measure \
