+++ pcli/Cargo.toml
@@ -37,6 +37,7 @@ penumbra-custody = { path = "../custody" }
 penumbra-tct = { path = "../tct" }
 penumbra-shielded-pool = { path = "../crates/shielded-pool", default-features = false }
 penumbra-ibc = { path = "../crates/ibc", default-features = false }
+penumbra-dao = { path = "../crates/dao", default-features = false }
 # TODO: replace by a penumbra-app
 penumbra-app = { path = "../app" }
