+++ transaction/Cargo.toml
@@ -12,6 +12,7 @@ penumbra-crypto = { path = "../crypto/" }
 penumbra-chain = { path = "../chain/", default-features = false }
 penumbra-shielded-pool = { path = "../crates/shielded-pool/", default-features = false }
 penumbra-ibc = { path = "../crates/ibc/", default-features = false }
+penumbra-dao = { path = "../crates/dao/", default-features = false }
 penumbra-tct = { path = "../tct" }
 penumbra-proof-params = { path = "../proof-params/", features = [
     "proving-keys",
