+++ transaction/src/plan/action.rs
@@ -1,4 +1,5 @@
 use penumbra_crypto::Balance;
+use penumbra_dao::{DaoDeposit, DaoOutput, DaoSpend};
 use penumbra_ibc::{IbcAction, Ics20Withdrawal};
 use penumbra_proto::{
     core::stake::v1alpha1 as pb_stake, core::transaction::v1alpha1 as pb_t, DomainType,
@@ -19,8 +20,8 @@ pub use swap_claim::SwapClaimPlan;
 pub use undelegate_claim::UndelegateClaimPlan;
 
 use crate::action::{
-    DaoDeposit, DaoOutput, DaoSpend, Delegate, PositionClose, PositionOpen, ProposalDepositClaim,
-    ProposalSubmit, ProposalWithdraw, Undelegate, ValidatorVote,
+    Delegate, PositionClose, PositionOpen, ProposalDepositClaim, ProposalSubmit, ProposalWithdraw,
+    Undelegate, ValidatorVote,
 };
 
 /// A declaration of a planned [`Action`], for use in transaction creation.
